refractor(models): seccionar models.py en client.py, user.py y admin.py
