+++ users/models/__init__.py
@@ -1,0 +1,4 @@
+from .user import Usuarios
+from .admin import Producto, ProductoImagen
+from .client import Carrito, Tarjeta, Direccion
+from .user import Compra, DetalleCompra
